+++ Assets/Clues/HudRuntimeBootstrapper.cs
@@ -1,0 +1,76 @@
+using UnityEngine;
+using UnityEngine.EventSystems;
+using UnityEngine.InputSystem.UI;
+using UnityEngine.UI;
+
+namespace EscapeRoom
+{
+    public static class HudRuntimeBootstrapper
+    {
+        [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterSceneLoad)]
+        private static void Bootstrap()
+        {
+            EnsureHudCanvas();
+            EnsureEventSystem();
+            EnsureRuntimeObject<ClueJournalManager>("ClueJournalManager");
+            EnsureRuntimeObject<StoryProgressManager>("StoryProgressManager");
+            EnsureRuntimeObject<ClueJournalUI>("ClueJournalUI");
+            EnsureRuntimeObject<TimerUI>("TimerUI");
+            EnsureRuntimeObject<SettingsUI>("SettingsUI");
+            EnsureRuntimeObject<CluePickupPopupUI>("CluePickupPopupUI");
+        }
+
+        private static void EnsureHudCanvas()
+        {
+            GameObject canvasObject = GameObject.Find("HUD_Canvas");
+            if (canvasObject == null)
+            {
+                canvasObject = new GameObject("HUD_Canvas");
+            }
+
+            Canvas canvas = canvasObject.GetComponent<Canvas>();
+            if (canvas == null)
+            {
+                canvas = canvasObject.AddComponent<Canvas>();
+            }
+
+            canvas.renderMode = RenderMode.ScreenSpaceOverlay;
+
+            if (canvasObject.GetComponent<CanvasScaler>() == null)
+            {
+                CanvasScaler scaler = canvasObject.AddComponent<CanvasScaler>();
+                scaler.uiScaleMode = CanvasScaler.ScaleMode.ScaleWithScreenSize;
+                scaler.referenceResolution = new Vector2(1920f, 1080f);
+                scaler.matchWidthOrHeight = 0.5f;
+            }
+
+            if (canvasObject.GetComponent<GraphicRaycaster>() == null)
+            {
+                canvasObject.AddComponent<GraphicRaycaster>();
+            }
+        }
+
+        private static void EnsureEventSystem()
+        {
+            if (Object.FindObjectOfType<EventSystem>() != null)
+            {
+                return;
+            }
+
+            GameObject eventSystemObject = new GameObject("EventSystem");
+            eventSystemObject.AddComponent<EventSystem>();
+            eventSystemObject.AddComponent<InputSystemUIInputModule>();
+        }
+
+        private static void EnsureRuntimeObject<T>(string objectName) where T : Component
+        {
+            if (Object.FindObjectOfType<T>() != null)
+            {
+                return;
+            }
+
+            GameObject go = new GameObject(objectName);
+            go.AddComponent<T>();
+        }
+    }
+}